requirements.txt
-i https://pypi.org/simple

cachecontrol==0.12.6

cachetools==4.2.2; python_version ~= '3.5'

certifi==2020.12.5

chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

click==8.0.0; python_version >= '3.6'

colorama==0.4.4; platform_system == 'Windows'

dnspython==2.1.0; python_version >= '3.6'

email-validator==1.1.2

flask-login==0.5.0

flask-mongoengine==1.0.0

flask-wtf==0.14.3

flask==2.0.0

google-auth==1.30.0

idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

itsdangerous==2.0.0; python_version >= '3.6'

jinja2==3.0.0; python_version >= '3.6'

markupsafe==2.0.0; python_version >= '3.6'

mongoengine==0.23.1

msgpack==1.0.2

pyasn1-modules==0.2.8

pyasn1==0.4.8

pymongo==3.11.4

python-dotenv==0.17.1

requests==2.25.1

rsa==4.7.2; python_version >= '3.6'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

werkzeug==2.0.0; python_version >= '3.6'

wtforms[email]==2.3.3

